time change requests accept deny

+++ mern_stack/routes/timeChangeRoutes.js
@@ -62,32 +62,38 @@ module.exports = function(app) {
 	});
 
 	app.post('/acceptTimeChangeRequest', async (req, res) => {
-		const { timeChangeRequest_id, newclockintime, newduration } = req.body;
+		const { timeChangeRequest_id} = req.body;
 
 		try {
 			const timechangerequest = await TimeChangeRequest.findById(timeChangeRequest_id);
-			const duration = timechangerequest.duration;
-			const durationms = ((duration.hours * 3600) + (duration.minutes * 60) + duration.seconds) * 1000;
-			const clockouttimems = timechangerequest.clock_in_time + durationms;
-			const newclockouttime = new Date(clockouttimems);
-
-			const update = {
-				duration: newduration,
-				clock_in_time: newclockintime,
-				clock_out_time: newclockouttime
+			if (timechangerequest)
+			{
+				const duration = timechangerequest.new_duration;
+				const durationms = ((duration.hours * 3600) + (duration.minutes * 60) + duration.seconds) * 1000;
+	
+				const clockouttimems = timechangerequest.new_clock_in_time.getTime() + durationms;
+	
+				const newclockouttime = new Date(clockouttimems);
+	
+				const update = {
+					duration: timechangerequest.new_duration,
+					clock_in_time: timechangerequest.new_clock_in_time,
+					clock_out_time: newclockouttime
+				}
+				await ClockInOut.findByIdAndUpdate(timechangerequest.clock_in_out, update);
+				await TimeChangeRequest.findByIdAndDelete(timeChangeRequest_id);
 			}
-			await ClockInOut.findByIdAndUpdate(timechangerequest.clock_in_out, update);
-			await TimeChangeRequest.findByIdAndDelete(timeChangeRequest_id);
 		} catch (err) {
 			console.error('Error fetching clockinouts:', err);
 		}
 	});
 
 	app.post('/denyTimeChangeRequest', async (req, res) => {
+		console.log("trying to decline");
 		const { timeChangeRequest_id } = req.body;
 
 		try {
-			TimeChangeRequest.findByIdAndDelete(timeChangeRequest_id);
+			await TimeChangeRequest.findByIdAndDelete(timeChangeRequest_id);
 		} catch (err) {
 			console.error('Error fetching clockinouts:', err);
 		}
